+++ e2e/pages/administration/system-maintenance/qms-standard/qms-standard.po.js
@@ -1,0 +1,14 @@
+import SystemMaintenancePage from '../system-maintenance.po';
+
+class QmsStandardPage extends SystemMaintenancePage {
+  constructor() {
+    super();
+    this.elements = {
+      ...this.elements,
+      dataTable: 'table[aria-label="QMS Standard table"]',
+      addButton: '#add-new-qms-standard',
+    };
+  }
+}
+
+export default QmsStandardPage;