+++ e2e/pages/organizations/developers/developers.po.js
@@ -7,6 +7,7 @@ const elements = {
   products: '.products__product',
   editProductName: '#product-name',
   editProductCode: '#product-code',
+  currentProductOwner: '#product-owner-current',
   activeVersion: '#active-version',
   editVersionName: '#version-name',
   editVersionCode: '#version-code',
@@ -20,7 +21,7 @@ const elements = {
   developerStatus: '#developer-status-0',
   splitDeveloper: '#developer-component-split',
   developerName: '#name',
-  mergeDeveloper: '#developer-component-merge',
+  joinDeveloper: '#developer-component-join',
   editWebsite: '#website',
   contact: {
     fullName: '#fullName',
@@ -97,6 +98,14 @@ class DevelopersPage {
 
   get editProductCode() {
     return $(this.elements.editProductCode);
+  }
+
+  get currentProductOwner() {
+    return $(this.elements.currentProductOwner);
+  }
+
+  selectProductOwner(developerName) {
+    this.currentProductOwner.selectByVisibleText(developerName);
   }
 
   get editVersionName() {
@@ -260,8 +269,8 @@ class DevelopersPage {
     return $(this.elements.developerName);
   }
 
-  get mergeDeveloper() {
-    return $(this.elements.mergeDeveloper);
+  get joinDeveloper() {
+    return $(this.elements.joinDeveloper);
   }
 
   get fullName() {
@@ -308,8 +317,8 @@ class DevelopersPage {
     $(`#developers-product-move-new-${id}`).click();
   }
 
-  moveDeveloperToBeMerged(developerName) {
-    $(`//div[text()='${developerName}']`).click();
+  moveDeveloperToBeJoined(developerName) {
+    $(`td=${developerName}`).parentElement().$('button').click();
   }
 
   selectAllCertificationStatus() {