+++ src/app/components/components.module.js
@@ -7,6 +7,8 @@ import {
   ChplAttestationEditWrapper,
   ChplAttestationsViewWrapper,
 } from './attestation';
+import ChplBrowserComparedWidgetWrapper from './browser/browser-compared-widget-wrapper';
+import ChplBrowserViewedWidgetWrapper from './browser/browser-viewed-widget-wrapper';
 import ChplChangeRequestsWrapper from './change-request/change-requests-wrapper';
 import ChplCmsDisplayWrapper from './cms-widget/cms-display-wrapper';
 import ChplCompareDisplayWrapper from './compare-widget/compare-display-wrapper';
@@ -19,10 +21,14 @@ import {
   ChplConfirmProgress,
   ChplConfirmVersion,
 } from './listing/confirm';
-import { ChplCriteria } from './listing/details/criteria';
+import ChplCqms from './listing/details/cqms/cqms-wrapper';
+import ChplCriteriaWrapper from './listing/details/criteria/criteria-wrapper';
+import ChplG1g2 from './listing/details/g1g2/g1g2-wrapper';
 import ChplIcsFamily from './listing/details/ics-family/ics-family-wrapper';
 import ChplSurveillanceView from './listing/details/surveillance/surveillance-view';
+import ChplSed from './listing/details/sed/sed-wrapper';
 import ChplSedDetailsEditWrapper from './listing/details/sed/details-edit/wrapper';
+import ChplSedTaskView from './listing/details/sed/sed-task-view-wrapper';
 import ChplComplaintsWrapper from './surveillance/complaints/complaints-wrapper';
 import { ChplSurveillanceActivityReportingDateSelector } from './surveillance/manage';
 import {
@@ -63,6 +69,8 @@ angular
   .component('chplAttestationCreateWrapperBridge', reactToAngularComponent(ChplAttestationCreateWrapper))
   .component('chplAttestationEditWrapperBridge', reactToAngularComponent(ChplAttestationEditWrapper))
   .component('chplAttestationsViewWrapperBridge', reactToAngularComponent(ChplAttestationsViewWrapper))
+  .component('chplBrowserComparedWidgetBridge', reactToAngularComponent(ChplBrowserComparedWidgetWrapper))
+  .component('chplBrowserViewedWidgetBridge', reactToAngularComponent(ChplBrowserViewedWidgetWrapper))
   .component('chplChangeRequestsWrapperBridge', reactToAngularComponent(ChplChangeRequestsWrapper))
   .component('chplComplaintsWrapperBridge', reactToAngularComponent(ChplComplaintsWrapper))
   .component('chplCmsDisplayWrapperBridge', reactToAngularComponent(ChplCmsDisplayWrapper))
@@ -73,14 +81,18 @@ angular
   .component('chplConfirmProgressBridge', reactToAngularComponent(ChplConfirmProgress))
   .component('chplConfirmVersionBridge', reactToAngularComponent(ChplConfirmVersion))
   .component('chplConfirmationBridge', reactToAngularComponent(ChplConfirmation))
-  .component('chplCriteriaBridge', reactToAngularComponent(ChplCriteria))
+  .component('chplCqmsBridge', reactToAngularComponent(ChplCqms))
+  .component('chplCriteriaBridge', reactToAngularComponent(ChplCriteriaWrapper))
   .component('chplDirectReviewsViewBridge', reactToAngularComponent(ChplDirectReviewsView))
   .component('chplDeveloperBridge', reactToAngularComponent(ChplDeveloper))
   .component('chplEllipsisBridge', reactToAngularComponent(ChplEllipsis))
+  .component('chplG1g2Bridge', reactToAngularComponent(ChplG1g2))
   .component('chplIcsFamilyBridge', reactToAngularComponent(ChplIcsFamily))
   .component('chplLinkBridge', reactToAngularComponent(ChplLink))
   .component('chplNonProdIndicatorBridge', reactToAngularComponent(ChplNonProdIndicator))
+  .component('chplSedBridge', reactToAngularComponent(ChplSed))
   .component('chplSedDetailsEditWrapperBridge', reactToAngularComponent(ChplSedDetailsEditWrapper))
+  .component('chplSedTaskViewBridge', reactToAngularComponent(ChplSedTaskView))
   .component('chplSurveillanceActivityReportingDateSelectorBridge', reactToAngularComponent(ChplSurveillanceActivityReportingDateSelector))
   .component('chplSurveillanceViewBridge', reactToAngularComponent(ChplSurveillanceView))
   .component('chplUploadPromotingInteroperabilityBridge', reactToAngularComponent(ChplUploadPromotingInteroperability))