+++ src/app/components/filter/filter-panel.jsx
@@ -22,7 +22,7 @@ const useStyles = makeStyles({
     color: '#000',
   },
   filterPanelContainer: {
-    background: '#fafdff',
+    background: '#fff',
     display: 'grid',
     gridTemplateColumns: '1fr',
     rowGap: '16px',
@@ -61,6 +61,7 @@ const useStyles = makeStyles({
     display: 'grid',
     gridTemplateColumns: '1fr',
     alignItems: 'center',
+    minHeight: '8vh',
     maxHeight: '40vh',
     overflowY: 'auto',
   },
@@ -194,13 +195,14 @@ function ChplFilterPanel() {
         }}
         PaperProps={{
           style: {
-            width: '95%',
+            width: '93%',
             alignItems: 'center',
             borderRadius: '0 0 8px 8px',
             marginTop: '20px',
+            marginLeft: '32px',
             border: `1px solid ${palette.grey}`,
             boxShadow: 'rgb(149 157 165 / 40%) 0px 6px 16px 6px',
-            backgroundColor: '#fafdff',
+            backgroundColor: '#fff',
           },
         }}
       >
@@ -306,7 +308,20 @@ function ChplFilterPanel() {
                 )}
               >
                 <div className={classes.filterGroupTwoContainer}>
-                  { activeCategory.getValueEntry({
+                  { activeCategory.disabled
+                    && (
+                      <>
+                        <Typography variant="body1" gutterBottom>
+                          This information is temporarily unavailable. Please check back later.
+                        </Typography>
+                        <Typography variant="body1">
+                          Surveillance and Direct Review information can be downloaded from the
+                          {' '}
+                          <a href="#/resources/download">Download the CHPL page</a>
+                        </Typography>
+                      </>
+                    )}
+                  { !activeCategory.disabled && activeCategory.getValueEntry({
                     filter: activeCategory,
                     handleFilterToggle,
                     handleFilterUpdate,