+++ src/app/components/listing/details/criteria/criterion-details-edit.jsx
@@ -1,7 +1,4 @@
-import React, { useState } from 'react';
-import { bool, func } from 'prop-types';
-import CheckOutlinedIcon from '@material-ui/icons/CheckOutlined';
-import CloseOutlinedIcon from '@material-ui/icons/CloseOutlined';
+import React, { useContext, useState } from 'react';
 import {
   Button,
   Card,
@@ -15,21 +12,24 @@ import {
   Typography,
   makeStyles,
 } from '@material-ui/core';
+import CheckOutlinedIcon from '@material-ui/icons/CheckOutlined';
+import CloseOutlinedIcon from '@material-ui/icons/CloseOutlined';
+import { bool, func } from 'prop-types';
 import { useFormik } from 'formik';
 import * as yup from 'yup';
 
-import { ChplReliedUponSoftwareEdit } from './relied-upon-software';
-import ChplConformanceMethodsEdit from './conformance-methods';
-import ChplOptionalStandardsEdit from './optional-standards';
-import ChplSvapsEdit from './svaps';
-import ChplFunctionalitiesTestedEdit from './functionalities-tested';
-import ChplTestDataEdit from './test-data';
-import ChplTestProceduresEdit from './test-procedures';
-import ChplTestStandardsEdit from './test-standards';
-import ChplTestToolsEdit from './test-tools';
+import ChplReliedUponSoftwareEdit from './relied-upon-software/relied-upon-software-edit';
+import ChplConformanceMethodsEdit from './conformance-methods/conformance-methods-edit';
+import ChplOptionalStandardsEdit from './optional-standards/optional-standards-edit';
+import ChplSvapsEdit from './svaps/svaps-edit';
+import ChplFunctionalitiesTestedEdit from './functionalities-tested/functionalities-tested-edit';
+import ChplTestDataEdit from './test-data/test-data-edit';
+import ChplTestProceduresEdit from './test-procedures/test-procedures-edit';
+import ChplTestStandardsEdit from './test-standards/test-standards-edit';
+import ChplTestToolsEdit from './test-tools/test-tools-edit';
 
 import { ChplTextField } from 'components/util';
-import { getAngularService } from 'services/angular-react-helper';
+import { UserContext } from 'shared/contexts';
 import {
   certificationResult,
   resources as resourceDefinition,
@@ -46,14 +46,14 @@ const validationSchema = yup.object({
     .url('Enter a valid URL'),
 });
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles({
   subtitleSpacing: {
     marginBottom: '8px',
   },
-}));
+});
 
 function ChplCriterionDetailsEdit(props) {
-  const { hasAnyRole } = getAngularService('authService');
+  const { hasAnyRole } = useContext(UserContext);
   const { hasIcs, isConfirming } = props;
   /* eslint-disable react/destructuring-assignment */
   const resources = {