+++ src/app/components/listing/details/criteria/functionalities-tested/index.js
@@ -1,3 +1,0 @@
-import ChplFunctionalitiesTestedEdit from './functionalities-tested-edit';
-
-export default ChplFunctionalitiesTestedEdit;