+++ src/app/components/listing/details/criteria/test-tools/index.js
@@ -1,3 +1,0 @@
-import ChplTestToolsEdit from './test-tools-edit';
-
-export default ChplTestToolsEdit;