+++ src/app/components/listing/details/sed/sed-wrapper.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import ChplSed from './sed';
+
+import AppWrapper from 'app-wrapper';
+import { listing as listingType } from 'shared/prop-types/listing';
+
+function ChplSedWrapper(props) {
+  const { listing } = props;
+
+  return (
+    <AppWrapper>
+      <ChplSed listing={listing} />
+    </AppWrapper>
+  );
+}
+
+export default ChplSedWrapper;
+
+ChplSedWrapper.propTypes = {
+  listing: listingType.isRequired,
+};